Synthesize the oscillator waveforms in f64, as the C++ does
`GetOneSample_Overtone` and `GetOneSample_Coodinate` are `double` throughout,
down to `double pi = 3.1415926535897932` and libm's `sin`:

    sss          = 2 * pi * ( _p_point[ o ].x ) * index / _sample_num;
    work_double += ( sin( sss ) * (double)_p_point[ o ].y / ( _p_point[ o ].x ) / 128 );

This port worked in `f32`, whose 1e-7 of headroom is not enough: the table entry
that comes out of it is truncated to an `i16`, so a value that lands near an
integer goes the other way. That was the last of the noise instruments.

`lite_math::sin_f64` hands back the `f64` the reduction and the polynomial
already produced instead of narrowing it, so `sin` is now that with a cast on
either side and its `f32` results are unchanged. The polynomial is good to 3e-9,
short of an `f64` ulp but two orders past what `f32` could carry, and enough here.

All 47 noise instruments now match the reference exactly, 48 of the 53 renders
in all, for 511 bytes of wasm. The five songs left are closer than they were --
`Aisatsu[Rusk]` from 211716 differing samples to 211690, `5LOVE[Quois]` from
143378 to 143032 -- since the PTV voices go through the same oscillator, but
what remains there is the mixing chain.

+++ src/woice.rs
@@ -661,7 +661,7 @@ fn update_wave_ptv(
     };
     let frame = &mut instance.samples[s as usize];
     for (c, &pv) in pan_vol.iter().enumerate().take(channels as usize) {
-      let raw = osc * pv as f32 / 64.0;
+      let raw = osc * pv as f64 / 64.0;
       if raw.abs() > 1.0 {
         clipped = true;
       }
